src/bot/requirements.txt
aiavatar==0.7.2
discord.py[voice]==2.3.2
sounddevice==0.4.6
numpy>=2.2.3
soundfile==0.12.1
python-dotenv==1.0.0
sqlalchemy==2.0.23 